pyproject.toml
[project]
name = "integracao-imox"
version = "0.1.0"
description = "Script for webhook integration with trigger activation to send data from Redcap to Polotrial."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
